backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
pillow>=10.0.0
python-decouple>=3.8
